Added tests and checks for command line usage of resolution setting
