+++ openmdao/visualization/meta_model_viewer/meta_model_visualization.py
@@ -894,6 +894,8 @@ def view_metamodel(meta_model_comp, resolution, port_number):
     ----------
     meta_model_comp : MetaModelStructuredComp or MetaModelUnStructuredComp
         The metamodel component.
+    resolution : int
+        Number of points to control contour plot resolution.
     port_number : int
         Bokeh plot port number.
     """
